+++ ent/apikey/where.go
@@ -9,52 +9,52 @@ import (
 )
 
 // ID filters vertices based on their ID field.
-func ID(id int64) predicate.APIKey {
+func ID(id int) predicate.APIKey {
 	return predicate.APIKey(sql.FieldEQ(FieldID, id))
 }
 
 // IDEQ applies the EQ predicate on the ID field.
-func IDEQ(id int64) predicate.APIKey {
+func IDEQ(id int) predicate.APIKey {
 	return predicate.APIKey(sql.FieldEQ(FieldID, id))
 }
 
 // IDNEQ applies the NEQ predicate on the ID field.
-func IDNEQ(id int64) predicate.APIKey {
+func IDNEQ(id int) predicate.APIKey {
 	return predicate.APIKey(sql.FieldNEQ(FieldID, id))
 }
 
 // IDIn applies the In predicate on the ID field.
-func IDIn(ids ...int64) predicate.APIKey {
+func IDIn(ids ...int) predicate.APIKey {
 	return predicate.APIKey(sql.FieldIn(FieldID, ids...))
 }
 
 // IDNotIn applies the NotIn predicate on the ID field.
-func IDNotIn(ids ...int64) predicate.APIKey {
+func IDNotIn(ids ...int) predicate.APIKey {
 	return predicate.APIKey(sql.FieldNotIn(FieldID, ids...))
 }
 
 // IDGT applies the GT predicate on the ID field.
-func IDGT(id int64) predicate.APIKey {
+func IDGT(id int) predicate.APIKey {
 	return predicate.APIKey(sql.FieldGT(FieldID, id))
 }
 
 // IDGTE applies the GTE predicate on the ID field.
-func IDGTE(id int64) predicate.APIKey {
+func IDGTE(id int) predicate.APIKey {
 	return predicate.APIKey(sql.FieldGTE(FieldID, id))
 }
 
 // IDLT applies the LT predicate on the ID field.
-func IDLT(id int64) predicate.APIKey {
+func IDLT(id int) predicate.APIKey {
 	return predicate.APIKey(sql.FieldLT(FieldID, id))
 }
 
 // IDLTE applies the LTE predicate on the ID field.
-func IDLTE(id int64) predicate.APIKey {
+func IDLTE(id int) predicate.APIKey {
 	return predicate.APIKey(sql.FieldLTE(FieldID, id))
 }
 
 // UserID applies equality check predicate on the "user_id" field. It's identical to UserIDEQ.
-func UserID(v int64) predicate.APIKey {
+func UserID(v int) predicate.APIKey {
 	return predicate.APIKey(sql.FieldEQ(FieldUserID, v))
 }
 
@@ -69,22 +69,22 @@ func Name(v string) predicate.APIKey {
 }
 
 // UserIDEQ applies the EQ predicate on the "user_id" field.
-func UserIDEQ(v int64) predicate.APIKey {
+func UserIDEQ(v int) predicate.APIKey {
 	return predicate.APIKey(sql.FieldEQ(FieldUserID, v))
 }
 
 // UserIDNEQ applies the NEQ predicate on the "user_id" field.
-func UserIDNEQ(v int64) predicate.APIKey {
+func UserIDNEQ(v int) predicate.APIKey {
 	return predicate.APIKey(sql.FieldNEQ(FieldUserID, v))
 }
 
 // UserIDIn applies the In predicate on the "user_id" field.
-func UserIDIn(vs ...int64) predicate.APIKey {
+func UserIDIn(vs ...int) predicate.APIKey {
 	return predicate.APIKey(sql.FieldIn(FieldUserID, vs...))
 }
 
 // UserIDNotIn applies the NotIn predicate on the "user_id" field.
-func UserIDNotIn(vs ...int64) predicate.APIKey {
+func UserIDNotIn(vs ...int) predicate.APIKey {
 	return predicate.APIKey(sql.FieldNotIn(FieldUserID, vs...))
 }
